댓글창 만듬

+++ src/pages/PostPage/Post.tsx
@@ -2,6 +2,9 @@ import { useNavigate, useParams } from "react-router-dom";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
 import { removePost } from "../../store/posts/posts.slice";
 import { toast } from "react-toastify";
+import { toggleApplyStudyModal } from "../../store/modal/modalSlice";
+import ApplyStudyModal from "../../components/Modal/ApplyStudyModal/ApplyStudyModal";
+import Comment from "./Comment/Comment";
 
 const Post = () => {
   const { posts } = useAppSelector((state) => state.posts);
@@ -9,32 +12,43 @@ const Post = () => {
   const post = { ...posts.filter((post) => post.id === id)[0] };
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
+  const { viewApplyStudyModal } = useAppSelector((state) => state.modal);
 
   const handleRemoveClick = (
     e: React.MouseEvent<HTMLButtonElement, MouseEvent>
   ) => {
     e.preventDefault();
-    dispatch(removePost({ id: post.id }));
-    navigate(`/`);
-    toast.info("모집 글이 삭제되었습니다.");
+    if (window.confirm("모집글을 삭제하시겠습니까?")) {
+      dispatch(removePost({ id: post.id }));
+      navigate(`/`);
+      toast.info("모집 글이 삭제되었습니다.");
+    }
   };
 
   const handleEditClick = () => {
     navigate(`/write/${post.id}`);
   };
 
+  const handleApplyClick = () => {
+    dispatch(toggleApplyStudyModal(true));
+  };
+
   return (
     <div className="w-full flex flex-col items-center justify-center">
+      {viewApplyStudyModal && <ApplyStudyModal />}
       <div className="w-full flex items-center justify-center mt-4">
         <h2 className="text-title font-semibold text-5xl">CosMo&apos;s</h2>
       </div>
       <div className="mt-8">
         <h2 className="text-white font-bold text-3xl">{post.title}</h2>
       </div>
-      <div className="w-3/5 mt-8 mb-6">
+      <div className="flex items-center justify-between w-3/5 mt-8 mb-6">
         <p className="font-semibold text-2xl text-white">
           모집인원: <span className="text-forestGreen">{post.recruit}</span>명
         </p>
+        <div className="bg-skyBlue py-2 px-3 rounded-lg">
+          <span className="text-xl font-medium">모집중</span>
+        </div>
       </div>
       <div className="w-3/5 my-4">
         <p className="font-semibold text-2xl text-white">스터디 정보:</p>
@@ -42,7 +56,10 @@ const Post = () => {
           <p className="p-3 font-medium">{post.content}</p>
         </div>
       </div>
-      <div className="bg-gold my-6 rounded-lg">
+      <div className="w-3/5">
+        <Comment />
+      </div>
+      <div onClick={handleApplyClick} className="bg-gold my-6 rounded-lg">
         <button className="px-4 py-3 text-navy font-semibold text-lg">
           신청하기
         </button>
